+++ test/dummy/app/controllers/blorgh/posts_controller.rb
@@ -2,6 +2,8 @@
 
 module Blorgh
   class PostsController < ActionController::Base
+    around_action :set_locale_from_url
+
     def index
       render plain: I18n.locale
     end