+++ test/dummy/app/controllers/dummy_without_around_action_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class DummyWithoutAroundActionController < ActionController::Base
+  def dummy
+    render plain: I18n.locale
+  end
+end